SB-24484 : fix :unwanted loggs removed,showpop moved to observation-util
