+++ src/app/client/src/app/modules/observation/guards/ml/ml.guard.ts
@@ -27,8 +27,6 @@ export class MlGuard implements CanActivate {
         return false
       }
     }).catch(error => {
-      console.log("error");
-      console.log(error);
       return false
     })
   }
